Fix Gen3 scores
The Gen3 scores are generated using the distances to the depot vertex.
These distances are needed to round to the nearest integer before
calculating the scores. However, in the previous versions, the scores
were generated without rounding the distances.

This update affects to 8 instances: a280, fnl4461, nrw1379, rat195,
rat575, rat783, rat99, tsp225.